+++ src/main/java/com/zerobase/babdeusilbun/controller/InquiryController.java
@@ -1,18 +1,17 @@
-package com.zerobase.babdeusilbun.inquiry.controller;
+package com.zerobase.babdeusilbun.controller;
 
 import static org.springframework.http.HttpStatus.*;
 
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto.DetailResponse;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto.ListResponse;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryImageDto;
-import com.zerobase.babdeusilbun.inquiry.service.InquiryService;
+import com.zerobase.babdeusilbun.dto.InquiryDto;
+import com.zerobase.babdeusilbun.dto.InquiryDto.DetailResponse;
+import com.zerobase.babdeusilbun.dto.InquiryDto.ListResponse;
+import com.zerobase.babdeusilbun.dto.InquiryImageDto;
+import com.zerobase.babdeusilbun.service.InquiryService;
 import com.zerobase.babdeusilbun.security.dto.CustomUserDetails;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
@@ -21,7 +20,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
