+++ src/main/java/com/zerobase/babdeusilbun/dto/InquiryImageDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.babdeusilbun.inquiry.dto;
+package com.zerobase.babdeusilbun.dto;
 
 import com.zerobase.babdeusilbun.domain.InquiryImage;
 import lombok.AccessLevel;
