+++ src/main/java/com/zerobase/babdeusilbun/dto/MeetingRequest.java
@@ -1,13 +1,7 @@
-package com.zerobase.babdeusilbun.meeting.dto;
+package com.zerobase.babdeusilbun.dto;
 
-import com.zerobase.babdeusilbun.domain.Meeting;
-import com.zerobase.babdeusilbun.domain.User;
-import com.zerobase.babdeusilbun.dto.DeliveryAddressDto;
-import com.zerobase.babdeusilbun.dto.MetAddressDto;
 import com.zerobase.babdeusilbun.enums.PurchaseType;
-import jakarta.persistence.criteria.CriteriaBuilder.In;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
