+++ src/test/java/com/zerobase/babdeusilbun/inquiry/service/InquiryServiceTest.java
@@ -13,18 +13,13 @@ import com.zerobase.babdeusilbun.domain.Inquiry;
 import com.zerobase.babdeusilbun.domain.InquiryImage;
 import com.zerobase.babdeusilbun.domain.User;
 import com.zerobase.babdeusilbun.enums.InquiryStatus;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto.DetailResponse;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto.ListResponse;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryDto.Request;
-import com.zerobase.babdeusilbun.inquiry.dto.InquiryImageDto;
-import com.zerobase.babdeusilbun.inquiry.service.impl.InquiryServiceImpl;
+import com.zerobase.babdeusilbun.dto.InquiryDto.Request;
+import com.zerobase.babdeusilbun.service.impl.InquiryServiceImpl;
 import com.zerobase.babdeusilbun.repository.InquiryImageRepository;
 import com.zerobase.babdeusilbun.repository.InquiryRepository;
 import com.zerobase.babdeusilbun.repository.UserRepository;
 import com.zerobase.babdeusilbun.security.dto.CustomUserDetails;
-import java.lang.reflect.Array;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
