+++ src/test/java/com/zerobase/babdeusilbun/meeting/service/MeetingServiceTest.java
@@ -28,10 +28,10 @@ import com.zerobase.babdeusilbun.dto.MetAddressDto;
 import com.zerobase.babdeusilbun.enums.PurchaseStatus;
 import com.zerobase.babdeusilbun.enums.PurchaseType;
 import com.zerobase.babdeusilbun.exception.CustomException;
-import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest;
-import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest.Create;
-import com.zerobase.babdeusilbun.meeting.scheduler.MeetingScheduler;
-import com.zerobase.babdeusilbun.meeting.service.impl.MeetingServiceImpl;
+import com.zerobase.babdeusilbun.dto.MeetingRequest;
+import com.zerobase.babdeusilbun.dto.MeetingRequest.Create;
+import com.zerobase.babdeusilbun.scheduler.MeetingScheduler;
+import com.zerobase.babdeusilbun.service.impl.MeetingServiceImpl;
 import com.zerobase.babdeusilbun.repository.custom.impl.CustomMeetingRepositoryImpl;
 import com.zerobase.babdeusilbun.repository.MeetingRepository;
 import com.zerobase.babdeusilbun.repository.PurchaseRepository;
